new teams scoring method

+++ src/main/java/com/soapboxrace/core/bo/FriendBO.java
@@ -122,6 +122,10 @@ public class FriendBO {
 			if (!teamToJoin.getOpenEntry()) {
 				Long teamleaderId = teamToJoin.getLeader().getPersonaId();
 				List<ReportEntity> teamInviteCheck = reportDAO.findTeamInvite(teamleaderId, personaId);
+				if (teamToJoin.getPlayersCount() >= 8) {
+					openFireSoapBoxCli.send(XmppChat.createSystemMessage("### This team is full."), personaId);
+					return null;
+				}
 				if (!teamInviteCheck.isEmpty() && teamToJoin.getPlayersCount() < 8) {
 					teamsBO.teamJoinIG(personaSender, teamToJoin);
 					reportDAO.deleteTeamInvite(teamleaderId, personaId);
